fix: address PR #72 review issues — security, type safety, file splits
- Fix path traversal vulnerability in gateway-ops.ts: add assertPathWithinWorkspace()
  guard for host-only listFiles/readFile operations (resolves + startsWith check)
- Add protocol-layer path validation in extended-handlers.ts: reject null bytes
  and '..' segments before paths reach the ops layer
- Replace 'as any' casts in gateway-history.ts with proper type guards using
  ChatToolCallMessage, ChatAssistantMessage, and ToolResultImage types
- Replace fragile single-slot _pendingListPath/_pendingReadPath in files store
  with queue-based correlation (_pendingListQueue/_pendingReadQueue) to handle
  concurrent in-flight requests safely
- Extract 300-line QR generator script from qr-page.ts (464 LOC) into
  qr-generator.ts, bringing qr-page.ts down to 146 LOC
- Switch static-files.ts from sync fs.readFileSync to async createReadStream
  to avoid blocking the event loop on every static file request

All packages pass typecheck (25/25).
